Keep the new unit tests off the AITER import path so CI can run them
The Pre Checkin unit-test job failed with pytest exit code 2 -- a COLLECTION
error, not test failures. It runs on a plain CPU runner with no AITER, and the
two tests added for the round-3 fixes imported their subjects from modules that
pull AITER at module scope (`aiter_mla.py:11`, `deepseek_v2.py:30`), so both
files errored during collection and took the whole run down.

Reproduced locally by blocking `aiter` on the meta path: "Interrupted: 2 errors
during collection", exit 2, matching CI.

`importorskip` would have turned the job green while leaving the tests unrun,
which is the complaint the last two reviews already made about this PR's test
surface. Moved the subjects instead, into modules that need nothing but the
config:

* `atom/model_ops/mla_geometry.py` -- `mla_kv_entry_dim`, `mla_qk_head_dim`,
  `aligned_index_cache_dim`. Pure integer arithmetic on config fields.
* `atom/models/moe_router.py` -- `moe_router_dtype`. Needs torch and the
  config, nothing else.

Both follow the reasoning already written down in `atom/model_ops/__init__.py`,
which resolves its attention frontend lazily for exactly this purpose.
`kimi_mla_gdn_attn` now imports the geometry helpers straight from the light
module rather than through `aiter_mla`, so no re-export is left dangling.

Verified with AITER blocked: both modules import with zero `aiter.*` in
sys.modules, and all 8 tests run and pass. Local full suite is unchanged at
4831 passed / 14 failed -- those 14 fail identically on plain origin/main.
black clean, diff-context ruff clean.
